Reorganize tests into fixture and consumer package.
