+++ test_pluginlib/test/utest.cpp
@@ -32,15 +32,15 @@
 #include <memory>
 #include <pluginlib/class_loader.hpp>  // NOLINT
 
-#include "./test_base.h"
+#include "test_base.h"
 
 TEST(PluginlibTest, unknownPlugin) {
-  pluginlib::ClassLoader<test_base::Fubar> test_loader("pluginlib", "test_base::Fubar");
+  pluginlib::ClassLoader<test_base::Fubar> test_loader("test_pluginlib_fixture", "test_base::Fubar");
   ASSERT_THROW(test_loader.createInstance("pluginlib/foobar"), pluginlib::LibraryLoadException);
 }
 
 TEST(PluginlibTest, misspelledPlugin) {
-  pluginlib::ClassLoader<test_base::Fubar> bad_test_loader("pluginlib", "test_base::Fuba");
+  pluginlib::ClassLoader<test_base::Fubar> bad_test_loader("test_pluginlib_fixture", "test_base::Fuba");
   ASSERT_THROW(bad_test_loader.createInstance("pluginlib/foo"), pluginlib::LibraryLoadException);
 }
 
@@ -51,15 +51,15 @@ TEST(PluginlibTest, invalidPackage) {
 }
 
 TEST(PluginlibTest, brokenPlugin) {
-  pluginlib::ClassLoader<test_base::Fubar> test_loader("pluginlib", "test_base::Fubar");
+  pluginlib::ClassLoader<test_base::Fubar> test_loader("test_pluginlib_fixture", "test_base::Fubar");
   ASSERT_THROW(test_loader.createInstance("pluginlib/none"), pluginlib::PluginlibException);
 }
 
 TEST(PluginlibTest, workingPlugin) {
-  pluginlib::ClassLoader<test_base::Fubar> test_loader("pluginlib", "test_base::Fubar");
+  pluginlib::ClassLoader<test_base::Fubar> test_loader("test_pluginlib_fixture", "test_base::Fubar");
 
   try {
-    std::shared_ptr<test_base::Fubar> foo = test_loader.createInstance("pluginlib/foo");
+    std::shared_ptr<test_base::Fubar> foo = test_loader.createInstance("test_pluginlib_fixture/foo");
     foo->initialize(10.0);
     EXPECT_EQ(100.0, foo->result());
   } catch (pluginlib::PluginlibException & ex) {
@@ -72,10 +72,10 @@ TEST(PluginlibTest, workingPlugin) {
 
 TEST(PluginlibTest, createUnmanagedInstanceAndUnloadLibrary) {
   ROS_INFO("Making the ClassLoader...");
-  pluginlib::ClassLoader<test_base::Fubar> pl("pluginlib", "test_base::Fubar");
+  pluginlib::ClassLoader<test_base::Fubar> pl("test_pluginlib_fixture", "test_base::Fubar");
 
   ROS_INFO("Instantiating plugin...");
-  test_base::Fubar * inst = pl.createUnmanagedInstance("pluginlib/foo");
+  test_base::Fubar * inst = pl.createUnmanagedInstance("test_pluginlib_fixture/foo");
 
   ROS_INFO("Deleting plugin...");
   delete inst;
