+++ test_pluginlib_fixture/test/test_plugins.h
@@ -27,9 +27,56 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <pluginlib/class_list_macros.hpp>
-#include "./test_base.h"
-#include "test_plugins.h"  // NOLINT
+#ifndef TEST_PLUGINS_H_
+#define TEST_PLUGINS_H_
 
-PLUGINLIB_EXPORT_CLASS(test_plugins::Foo, test_base::Fubar)
-PLUGINLIB_EXPORT_CLASS(test_plugins::Bar, test_base::Fubar)
+#include <cmath>
+
+#include <test_pluginlib_fixture/test_base.h>
+
+namespace test_plugins
+{
+class Bar : public test_base::Fubar
+{
+public:
+  Bar() {}
+
+  void initialize(double foo)
+  {
+    foo_ = foo;
+  }
+
+  double result()
+  {
+    return 0.5 * foo_ * getBar();
+  }
+
+  double getBar()
+  {
+    return sqrt((foo_ * foo_) - ((foo_ / 2) * (foo_ / 2)));
+  }
+
+private:
+  double foo_;
+};
+
+class Foo : public test_base::Fubar
+{
+public:
+  Foo() {}
+
+  void initialize(double foo)
+  {
+    foo_ = foo;
+  }
+
+  double result()
+  {
+    return foo_ * foo_;
+  }
+
+private:
+  double foo_;
+};
+}  // namespace test_plugins
+#endif  // TEST_PLUGINS_H_
